refactor: 패키지와 파일명 수정
- Dto 대소문자 변경
- ServOrdericeTest 번호 지정
